+++ gpt4free/openaihosted/readme.md
@@ -2,7 +2,7 @@
 
 ```python
 # import library
-from openai_rev import openaihosted
+from gpt4free import openaihosted
 
 res = openaihosted.Completion.create(systemprompt="U are ChatGPT", text="What is 4+4",
                                      assistantprompt="U are a helpful assistant.").text
