+++ gpt4free/quora/graphql/DeleteMessageMutation.graphql
