+++ gpt4free/theb/theb_test.py
@@ -1,4 +1,4 @@
-from openai_rev import theb
+from gpt4free import theb
 
 for token in theb.Completion.create('hello world'):
     print(token, end='', flush=True)
