+++ testing/test_main.py
@@ -1,29 +1,29 @@
-import openai_rev
-from openai_rev import Provider, quora, forefront
+import gpt4free
+from gpt4free import Provider, quora, forefront
 
 # usage You
-response = openai_rev.Completion.create(Provider.You, prompt='Write a poem on Lionel Messi')
+response = gpt4free.Completion.create(Provider.You, prompt='Write a poem on Lionel Messi')
 print(response)
 
 # usage Poe
 token = quora.Account.create(logging=False)
-response = openai_rev.Completion.create(
+response = gpt4free.Completion.create(
     Provider.Poe, prompt='Write a poem on Lionel Messi', token=token, model='ChatGPT'
 )
 print(response)
 
 # usage forefront
 token = forefront.Account.create(logging=False)
-response = openai_rev.Completion.create(
+response = gpt4free.Completion.create(
     Provider.ForeFront, prompt='Write a poem on Lionel Messi', model='gpt-4', token=token
 )
 print(response)
 print(f'END')
 
 # usage theb
-response = openai_rev.Completion.create(Provider.Theb, prompt='Write a poem on Lionel Messi')
+response = gpt4free.Completion.create(Provider.Theb, prompt='Write a poem on Lionel Messi')
 print(response)
 
 # usage openaihosted
-response = openai_rev.Completion.create(Provider.OpenAiHosted, prompt='Write a poem on Lionel Messi')
+response = gpt4free.Completion.create(Provider.OpenAiHosted, prompt='Write a poem on Lionel Messi')
 print(response)
