+++ testing/you_test.py
@@ -1,4 +1,4 @@
-from openai_rev import you
+from gpt4free import you
 
 # simple request with links and details
 response = you.Completion.create(prompt="hello world", detailed=True, include_links=True)
